requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
autoflake==0.7
autopep8==1.3.2
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
coverage==4.4.1           # via coveralls, pytest-cov
coveralls==1.2.0
django-autoslug==1.9.3
django-easy-select2==1.3.4
django-enumfields==0.9.0
django-environ==0.4.4
django-filter==1.0.4
django-parler-rest==1.4.2
django-parler==1.8        # via django-parler-rest
django==1.11.6            # via django-environ
djangorestframework-gis==0.11.2
djangorestframework==3.6.4
docopt==0.6.2             # via coveralls
factory-boy==2.9.2        # via pytest-factoryboy
faker==0.8.4              # via factory-boy
first==2.0.1              # via pip-tools
flake8-isort==2.2.2
flake8-polyfill==1.0.1    # via flake8-isort
flake8==3.4.1
httpio==0.1.3
idna==2.6                 # via requests
inflection==0.3.1         # via pytest-factoryboy
isort==4.2.15             # via flake8-isort
jsonschema==2.6.0
lxml==4.0.0
mccabe==0.6.1             # via flake8
pip-tools==1.10.1
psycopg2==2.7.3.1
py==1.4.34                # via pytest
pycodestyle==2.3.1        # via autopep8, flake8
pyflakes==1.5.0           # via autoflake, flake8
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-factoryboy==1.3.1
pytest==3.2.3             # via pytest-cov, pytest-django, pytest-factoryboy
python-dateutil==2.6.1    # via faker
pytz==2017.2
pyyaml==3.12
requests==2.18.4          # via coveralls, httpio
six==1.11.0               # via django-environ, faker, pip-tools, python-dateutil
testfixtures==5.2.0       # via flake8-isort
urllib3==1.22             # via requests
